requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    make upgrade-requirements
#
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via openedx-events
cffi==1.16.0
    # via pynacl
click==8.1.7
    # via edx-django-utils
django==3.2
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/package.in
    #   django-config-models
    #   django-crum
    #   django-extensions
    #   django-waffle
    #   djangorestframework
    #   edx-django-utils
    #   openedx-events
django-config-models==2.5.1
    # via -r requirements/package.in
django-crum==0.7.9
    # via edx-django-utils
django-extensions==3.2.3
    # via -r requirements/package.in
django-waffle==4.1.0
    # via edx-django-utils
djangorestframework==3.14.0
    # via django-config-models
edx-django-utils==5.10.1
    # via
    #   django-config-models
    #   openedx-events
edx-opaque-keys[django]==2.5.1
    # via openedx-events
fastavro==1.9.4
    # via openedx-events
newrelic==9.7.0
    # via edx-django-utils
openedx-events==9.5.2
    # via -r requirements/package.in
pbr==6.0.0
    # via stevedore
psutil==5.9.8
    # via edx-django-utils
pycparser==2.21
    # via cffi
pymongo==3.13.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pytz==2024.1
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.4
    # via django
stevedore==5.2.0
    # via
    #   edx-django-utils
    #   edx-opaque-keys
typing-extensions==4.9.0
    # via
    #   asgiref
    #   edx-opaque-keys
